Step

+++ server/index.js
@@ -1,3 +1,11 @@
+/**
+ * TODO
+ * разделить входящие сообщения на сценарии
+ * добавить коментарии
+ * добавить логирование
+ * добавить запись логирования
+ */
+
 // vendor
 const app = require('express');
 const http = require('http').Server(app);
@@ -7,15 +15,23 @@ const log = require('../libs/log');
 // configs
 const runes = require('./configs/runes');
 const server = require('./configs/server');
+
 const crypto = require('crypto');
+const SeedRandom = require('seedrandom');
 
 // classes
 const Changes = require('./classes/changes');
 const Board = require('./classes/board');
 const Lobby = require('./classes/lobby');
+const Step = require('./classes/step');
+const Player = require('./classes/player');
 
 const lobby = new Lobby();
 const game = {};
+
+function isGame(id) {
+  return Object.hasOwnProperty.call(game, id);
+}
 
 io.on('connection', (socket) => {
   const userName = socket.id.slice(0, 5);
@@ -25,22 +41,33 @@ io.on('connection', (socket) => {
     log('client', `[→] msg: ${msg}`);
   });
 
-  socket.on('lobby/ready', () => {
-    lobby.addPlayer(socket);
+  socket.on('lobby/ready', (name) => {
+    lobby.addPlayer(socket, name);
     if (lobby.isThereAnEnemy()) {
       //
       const id = crypto.randomBytes(4).toString('hex').toUpperCase();
 
+      const players = lobby.pairOfPlayers();
+      players[0].socket.join(id);
+      players[1].socket.join(id);
+
       game[id] = {
         changes: new Changes(),
-        board: new Board(runes, id),
+        board: new Board(runes),
+        players: [
+          new Player(players[0].name),
+          new Player(players[1].name),
+        ],
+        step: new Step(players),
+        seedRandom: new SeedRandom(id),
       };
 
-      const players = lobby.pairOfPlayers();
-      players[0].join(id);
-      players[1].join(id);
-
-      game[id].changes.add('loadBoard', { id, newBoard: game[id].board.generationBoard() });
+      game[id].changes.add('loadGame', {
+        gameID: id,
+        newBoard: game[id].board.generationBoard(game[id].seedRandom),
+        players: game[id].players,
+        step: game[id].step.coinToss(game[id].seedRandom),
+      });
       io.to(id).emit('changes', game[id].changes.release());
       //
     } else {
@@ -48,11 +75,15 @@ io.on('connection', (socket) => {
     }
   });
 
-
-  socket.on('game/reconnect', (id) => {
-    if (Object.hasOwnProperty.call(game, id)) {
+  socket.on('game/connect', (id) => {
+    if (isGame(id)) {
       socket.join(id);
-      game[id].changes.add('loadBoard', { id, newBoard: game[id].board.getBoard() });
+      game[id].changes.add('loadGame', {
+        gameID: id,
+        newBoard: game[id].board.getBoard(),
+        players: game[id].players,
+        step: game[id].step.getStep(),
+      });
       socket.emit('changes', game[id].changes.release());
     } else {
       socket.emit('changes', [{ event: 'noGame', data: 0 }]);
@@ -66,31 +97,36 @@ io.on('connection', (socket) => {
 
   // GAME
   socket.on('board/suggestion', (id) => {
-    game[id].changes.add('showSuggestion', game[id].board.findMoves());
-    socket.emit('changes', game[id].changes.release());
+    if (isGame(id)) {
+      game[id].changes.add('showSuggestion', game[id].board.findMoves());
+      socket.emit('changes', game[id].changes.release());
+    }
   });
 
-  socket.on('board/swap', (id, coordOne, coordTwo) => {
-    if (!Board.isEqualCoords(coordOne, coordTwo) &&
-         Board.isAdjacent(coordOne, coordTwo)) {
-      game[id].changes.add('swapRune', game[id].board.swap(coordOne, coordTwo));
-      game[id].board.findClusters(coordOne);
-      game[id].board.findClusters(coordTwo);
-      if (game[id].board.isClusters()) {
-        do {
-          game[id].changes.add('deleteRune', game[id].board.deleteClusters());
-          game[id].changes.add('dropRunes', game[id].board.drop());
-          game[id].changes.add('refillBoard', game[id].board.refill());
-          game[id].board.findAllClusters();
-        } while (game[id].board.isClusters());
-      } else {
-        game[id].changes.add('swapRune', game[id].board.swap(coordOne, coordTwo));
-        game[id].board.cleanClusters();
+  socket.on('board/swap', (id, name, coordOne, coordTwo) => {
+    if (isGame(id)) {
+      if (game[id].step.isStep(name)) {
+        if (!Board.isEqualCoords(coordOne, coordTwo) &&
+             Board.isAdjacent(coordOne, coordTwo)) {
+          game[id].changes.add('swapRune', game[id].board.swap(coordOne, coordTwo));
+          game[id].board.findClusters(coordOne);
+          game[id].board.findClusters(coordTwo);
+          if (game[id].board.isClusters()) {
+            do {
+              game[id].changes.add('deleteRune', game[id].board.deleteClusters());
+              game[id].changes.add('dropRunes', game[id].board.drop());
+              game[id].changes.add('refillBoard', game[id].board.refill(game[id].seedRandom));
+              game[id].board.findAllClusters();
+            } while (game[id].board.isClusters());
+            game[id].changes.add('nextStep', game[id].step.nextStep());
+          } else {
+            game[id].changes.add('swapRune', game[id].board.swap(coordOne, coordTwo));
+            game[id].board.cleanClusters();
+          }
+          io.to(id).emit('changes', game[id].changes.release());
+        }
       }
-    } else {
-      io.to(id).emit('msg', 'error');
     }
-    io.to(id).emit('changes', game[id].changes.release());
   });
 });
